moved algorithm into algo subdirectory
