+++ arrows/sfm/CMakeLists.txt
@@ -1,7 +1,8 @@
 # Build / Install Structure-from-Motion arrow
 
 set( plugin_sfm_headers
-  initialize_cameras_landmarks.h
+  algo/initialize_cameras_landmarks.h
+
   sfm_utils.h
   )
 
@@ -16,7 +17,8 @@ kwiver_install_headers(
   )
 
 set( plugin_sfm_sources
-  initialize_cameras_landmarks.cxx
+  algo/initialize_cameras_landmarks.cxx
+
   sfm_utils.cxx
   )
