+++ arrows/sfm/algo/initialize_cameras_landmarks.cxx
