+++ arrows/sfm/algo/initialize_cameras_landmarks.h
